Cleanup and organize a lot of Loader code.
